Use EVP interface for fallback
The use of standard EVP functions is preferred over the low level
interfaces because it will ensure the use of platform specific
acceleration, such as optimized assembly versions.

The drawback is increased use of memory, since we are leaving two open
CTXs (one for encryption, one for decryption) for every supported
cipher.  The contexts must be opened before the engine is registered;
otherwise, upon initializing the contexts, our own versions of the
cipher would be used instead of the standard openssl versions, and we
would open a new context and loop until there were too many open files.

To overcome this, we generate and initialize the contexts before the
engine is registered, and call the EVP_CIPHER init_key method directly.
It may not work reliably across all setups, but appears to work with the
internal openssl functions.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>