Use the CIPHERS/DIGESTS cmds to set fb thresholds
Use a colon and the threshold after the algorithm name in CIPHERS or
DIGESTS commands to set the fallback threshold.  This allows the
selection of a different threshold for every algorithm.

Strip the colon from the cipher list even when compiling with
AFALG_NO_FALLBACK, to keep the configuration file compatible with both
build options.

Signed-off-by: Eneas U de Queiroz <cotequeiroz@gmail.com>